avniproject/avni-webapp#1754 | Fix media table address_id type from numeric to integer to enable hash join with address table
The type mismatch forced a nested loop over 672K address rows per media row (19s).
Changing to integer allows a hash join (<1ms). AlterColumnType diff auto-migrates
existing databases on next ETL run instead of throwing on type mismatch.

+++ src/main/java/org/avniproject/etl/repository/rowMappers/tableMappers/MediaTable.java
@@ -17,7 +17,7 @@ public class MediaTable extends Table {
         return new Columns()
                 .withSerialIdColumn()
                 .withCommonColumns()
-                .withColumn(new Column("address_id", Column.Type.numeric, Column.ColumnType.index))
+                .withColumn(new Column("address_id", Column.Type.integer, Column.ColumnType.index))
                 .withColumn(new Column("image_url", Column.Type.text))
                 .withColumn(new Column("sync_parameter_key1", Column.Type.text))
                 .withColumn(new Column("sync_parameter_value1", Column.Type.text))
